Refactor duplicated serialer save logic.
